Fix new codechecker errors from clang10

+++ Detectors/MUON/MCH/Clustering/src/ClusterFinderOriginal.cxx
@@ -173,7 +173,7 @@ void ClusterFinderOriginal::resetPreCluster(gsl::span<const Digit>& digits)
     double y = mSegmentation->padPositionY(padID);
     double dx = mSegmentation->padSizeX(padID) / 2.;
     double dy = mSegmentation->padSizeY(padID) / 2.;
-    double charge = static_cast<double>(digit.getADC()) / std::numeric_limits<unsigned long>::max() * 1024;
+    double charge = static_cast<double>(digit.getADC()) / static_cast<double>(std::numeric_limits<unsigned long>::max()) * 1024;
     bool isSaturated = digit.getTime().time > 0;
     int plane = mSegmentation->isBendingPad(padID) ? 0 : 1;
